fix: wrapping up mazes with addition of training board
Set up a training board for monitoring agent's position in the gridworld, as well as current Q estimates and prioritized states. Fixed randmax function and tried Dyna with prioritized sweeping on a couple of hard mazes. Algorithm is not parallelized, and thus is quite slow on big mazes. However, the planning procedure is sensible and could be investigated further.